feat: port no-unused-vars rule
