+++ internal/rules/all.go
@@ -116,6 +116,7 @@ import (
 	"github.com/web-infra-dev/rslint/internal/rules/no_unused_expressions"
 	"github.com/web-infra-dev/rslint/internal/rules/no_unused_labels"
 	"github.com/web-infra-dev/rslint/internal/rules/no_unused_private_class_members"
+	"github.com/web-infra-dev/rslint/internal/rules/no_unused_vars"
 	"github.com/web-infra-dev/rslint/internal/rules/no_useless_backreference"
 	"github.com/web-infra-dev/rslint/internal/rules/no_useless_call"
 	"github.com/web-infra-dev/rslint/internal/rules/no_useless_catch"
@@ -279,6 +280,7 @@ func GetAllRules() []rule.Rule {
 		no_unused_expressions.NoUnusedExpressionsRule,
 		no_unused_labels.NoUnusedLabelsRule,
 		no_unused_private_class_members.NoUnusedPrivateClassMembersRule,
+		no_unused_vars.NoUnusedVarsRule,
 		one_var.OneVarRule,
 		prefer_arrow_callback.PreferArrowCallbackRule,
 		no_dupe_else_if.NoDupeElseIfRule,
