+++ internal/rules/no_unused_vars/no-unused-vars.schema.json
@@ -1,0 +1,53 @@
+{
+  "type": "array",
+  "items": [
+    {
+      "oneOf": [
+        {
+          "enum": ["all", "local"]
+        },
+        {
+          "type": "object",
+          "properties": {
+            "vars": {
+              "enum": ["all", "local"]
+            },
+            "varsIgnorePattern": {
+              "type": "string"
+            },
+            "args": {
+              "enum": ["all", "after-used", "none"]
+            },
+            "ignoreRestSiblings": {
+              "type": "boolean"
+            },
+            "argsIgnorePattern": {
+              "type": "string"
+            },
+            "caughtErrors": {
+              "enum": ["all", "none"]
+            },
+            "caughtErrorsIgnorePattern": {
+              "type": "string"
+            },
+            "destructuredArrayIgnorePattern": {
+              "type": "string"
+            },
+            "ignoreClassWithStaticInitBlock": {
+              "type": "boolean"
+            },
+            "ignoreUsingDeclarations": {
+              "type": "boolean"
+            },
+            "reportUsedIgnorePattern": {
+              "type": "boolean"
+            }
+          },
+          "additionalProperties": false
+        }
+      ]
+    }
+  ],
+  "minItems": 0,
+  "maxItems": 1
+}
