+++ packages/rslint-test-tools/rstest.config.mts
@@ -110,6 +110,7 @@ export default defineConfig({
     './tests/eslint/rules/no-undef.test.ts',
     './tests/eslint/rules/no-undef-init.test.ts',
     './tests/eslint/rules/no-unassigned-vars.test.ts',
+    './tests/eslint/rules/no-unused-vars.test.ts',
     './tests/eslint/rules/no-unused-private-class-members.test.ts',
     './tests/eslint/rules/no-var.test.ts',
     './tests/eslint/rules/one-var.test.ts',
